Incorporates libcpp __BASE_FILE__ fix.

gcc/upc/
	* upc-genericize.c (upc_expand_put): Use is_gimple_reg instead of
	is_gimple_non_addressable.

libgupc/
	* smp/upc_pgm_info.c (__upc_validate_pgm_info):  Ignore intervening
	null bytes in the UPC program information section, to ensure that
	all configuration information strings are processed.  Allow
	static THREADS compilations to be mixed with dynamic THREADS
	compilations as long as the static THREADS values all match.

From-SVN: r183075

+++ libgo/runtime/go-type-string.c
@@ -11,14 +11,14 @@
 
 /* A string hash function for a map.  */
 
-size_t
+uintptr_t
 __go_type_hash_string (const void *vkey,
-		       size_t key_size __attribute__ ((unused)))
+		       uintptr_t key_size __attribute__ ((unused)))
 {
-  size_t ret;
+  uintptr_t ret;
   const struct __go_string *key;
-  size_t len;
-  size_t i;
+  int len;
+  int i;
   const unsigned char *p;
 
   ret = 5381;
@@ -33,7 +33,7 @@ __go_type_hash_string (const void *vkey,
 
 _Bool
 __go_type_equal_string (const void *vk1, const void *vk2,
-			size_t key_size __attribute__ ((unused)))
+			uintptr_t key_size __attribute__ ((unused)))
 {
   const struct __go_string *k1;
   const struct __go_string *k2;
